CDE-main/Mudyeblog/requirements.txt
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
colorama==0.4.3
cycler==0.10.0
DateTime==4.3
dicttoxml==1.7.4
Django==3.1.2
django-birthday==0.1.3
django-urls==1.1.3
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
idna==3.1
is-disposable-email==1.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysql-connector-python==8.0.22
numpy==1.18.2
package-name==0.1
pefile==2019.4.18
Pillow==8.1.1
protobuf==3.13.0
pycparser==2.20
PyInstaller==3.6
pylint==2.5.3
PyMySQL==0.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.14.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-validators==1.0.0
xmltodict==0.12.0
zope.interface==5.2.0
